@@ -1,19 +1,48 @@
 [tool.poetry]
-name = "tc_c2pa_py"
-version = "0.1.0"
+name = "c2pie"
+version = "0.2.0"
 description = "The TC-C2PA-Py is python library which provide C2PA standard functionality."
-authors = ["Starodubtcev Dmitriy"]
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Operating System :: OS Independent",
+]
+authors = ["Tourmaline Core <contact@tourmalinecore.com>"]
+
+[project]
+name = "c2pie"
+authors = [ 
+  { name="Tourmaline Core", email="contact@tourmalinecore.com" },
+]
+
+[project.urls]
+Homepage = "https://github.com/TourmalineCore/tc-c2pa-py"
+Issues = "https://github.com/TourmalineCore/tc-c2pa-py/issues"
 
 [tool.poetry.dependencies]
-python = "^3.8"
-pytest = "^7.4.0"
+python = "^3.9,<3.13"
 cbor2 = "^5.6.3"
 cryptography = "^42.0.5"
 pytz = "^2024.1"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
+twine = "^6.1.0"
+pytest-cov = "^7.0.0"
 
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
+
+[tool.pytest.ini_options]
+markers = [
+    "e2e: marks tests as e2e",
+]
+
+[tool.coverage.run]
+omit = [
+  "*_test.py",
+  "*/__init__.py",
+]
+
+[tool.coverage.report]
+show_missing = true
